Fix issue with parallax being one frame behind
Summary:
The camera's FollowController updates the position to match the character, while the ParallaxController updates the backgrounds' positions to match the camera, leading to the backgrounds being one update behind.

Fix:
By moving the parallax update to LateUpdate, it will be able to use the updated camera position during the same update cycle.